Add Status Code and Error handling for the services, and controllers

+++ trempApplication/Properties/Services/CarService.cs
@@ -22,16 +22,17 @@ namespace trempApplication.Properties.Services
            
 
         }
-        public async Task AddCar(Car car)
+        public async Task<(bool IsSuccess, string ErrorMessage)> AddCar(Car car)
         {
             try
             {
                 if (car == null)
                 {
-                    throw new ArgumentNullException(nameof(car), "The car object is null.");
+                    return (false, "Car is null object");
                 }
 
                 await carsCollection.InsertOneAsync(car);
+                return(true,null);
             }
             catch (Exception ex)
             {
@@ -40,20 +41,23 @@ namespace trempApplication.Properties.Services
         }
 
 
-        public async Task DeleteCar(Guid carId)
+        public async Task<(bool IsSuccess, string ErrorMessage)> DeleteCar(Guid carId)
         {
             try
             {
                 var filter = Builders<Car>.Filter.Eq(u => u.Id, carId);
                 var car = await carsCollection.Find(filter).FirstOrDefaultAsync();
-
+                if(car == null)
+                {
+                    return (false, "No car was found to be deleted");
+                }
                 // Delete the car
                 await carsCollection.DeleteOneAsync(filter);
 
                 // Update the carIds property of all passengers that own the car
                 var update = Builders<Passenger>.Update.Pull(x => x.CarIds, carId);
                 var result = await passengersCollection.UpdateManyAsync(p => p.CarIds.Contains(carId), update);
-
+                return (true, null);
             }
             catch (Exception ex)
             {
@@ -64,12 +68,16 @@ namespace trempApplication.Properties.Services
 
 
 
-        public async Task<List<Car>> GetAllCars()
+        public async Task<(bool IsSuccess, List<Car> Car, string ErrorMessage)> GetAllCars()
         {
             try
             {
                 var cars = await carsCollection.Find(u => true).ToListAsync();
-                return cars;
+                if(cars != null)
+                {
+                    return (true, cars, null); 
+                }
+                return (false, null, "No cars found");
             }
             catch (Exception ex)
             {
@@ -78,13 +86,17 @@ namespace trempApplication.Properties.Services
             }
         }
 
-        public async Task<Car> GetCarById(Guid id)
+        public async Task<(bool IsSuccess, Car Car, string ErrorMessage)> GetCarById(Guid id)
         {
             try
             {
                 var filter = Builders<Car>.Filter.Eq(u => u.Id, id);
                 var car = await carsCollection.Find(filter).FirstOrDefaultAsync();
-                return car;
+                if(car != null)
+                {
+                    return(true, car, null);
+                }
+                return (false, null, "No car found");
             }
             catch (Exception ex)
             {
@@ -93,17 +105,18 @@ namespace trempApplication.Properties.Services
             }
         }
 
-        public async Task UpdateCar(Car car, Guid id)
+        public async Task<(bool IsSuccess, string ErrorMessage)> UpdateCar(Car car, Guid id)
         {
             try
             {
                 if (car == null)
                 {
-                    throw new ArgumentNullException(nameof(car), "The car object is null.");
+                    return(false, "The car object is null.");
                 }
 
                 var filter = Builders<Car>.Filter.Eq(u => u.Id, id);
                 await carsCollection.ReplaceOneAsync(filter, car);
+                return (true, null);
             }
             catch (Exception ex)
             {
